age calculator

+++ functions/calculator.py
@@ -1,3 +1,3 @@
 age = input('How old are you (in years) ?\n')
-age_in_days = int(age) * 365
+age_in_days = float(age) * 365
 print('Your are ' + str(age_in_days) + ' days old !!!')
